LayoutTests/fast/js/primitive-property-access-edge-cases-expected.txt

This page tests for assertion failures in edge cases of property lookup on primitive values.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS checkGet(1, Number) is true
PASS checkGet('hello', String) is true
PASS checkGet(true, Boolean) is true
PASS checkSet(1, Number) is true
PASS checkSet('hello', String) is true
PASS checkSet(true, Boolean) is true
PASS checkGetStrict(1, Number) is true
PASS checkGetStrict('hello', String) is true
PASS checkGetStrict(true, Boolean) is true
PASS checkSetStrict(1, Number) is true
PASS checkSetStrict('hello', String) is true
PASS checkSetStrict(true, Boolean) is true
PASS checkRead(1, Number) is true
PASS checkRead('hello', String) is true
PASS checkRead(true, Boolean) is true
PASS checkWrite(1, Number) is true
PASS checkWrite('hello', String) is true
PASS checkWrite(true, Boolean) is true
PASS checkReadStrict(1, Number) is true
PASS checkReadStrict('hello', String) is true
PASS checkReadStrict(true, Boolean) is true
PASS checkWriteStrict(1, Number) threw exception TypeError: Attempted to assign to readonly property..
PASS checkWriteStrict('hello', String) threw exception TypeError: Attempted to assign to readonly property..
PASS checkWriteStrict(true, Boolean) threw exception TypeError: Attempted to assign to readonly property..
PASS didNotCrash is true
PASS successfullyParsed is true

TEST COMPLETE
